Fixed Mac OS Target compilation errors

+++ MKNetworkKit/Categories/NSString+MKNetworkKitAdditions.m
@@ -32,7 +32,7 @@
 {
     const char *cStr = [self UTF8String];
     unsigned char result[16];
-    CC_MD5( cStr, strlen(cStr), result);
+    CC_MD5( cStr, (unsigned int) strlen(cStr), result);
     return [NSString stringWithFormat:
 			@"%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x",
 			result[0], result[1], result[2], result[3],
